+++ docs/changes.rst
@@ -14,6 +14,10 @@ Changelog
   Ref: scrum-1852
   [reinhardt]
 
+- Changed PDF generation engine to weasyprint
+  Ref: scrum-732
+  [reinhardt]
+
 
 15.2.0 (2024-01-08)
 -------------------