+++ docs/changes.rst
@@ -4,6 +4,9 @@ Changelog
 15.2.1 (unreleased)
 -------------------
 
+- Allow the client skin layer to be customized by other packages.
+  [ale-rt]
+
 - Allow the company model to be customized by other packages.
   [ale-rt]